Move Increment to Dimmer Funcs; Create Derived On for Dimmers
