+++ insteon_mqtt/device/functions/__init__.py
@@ -22,4 +22,4 @@ from .Scene import Scene
 from .SetAndState import SetAndState
 from .State import State
 from .Backlight import Backlight
-from .DimmerFlags import DimmerFlags
+from .DimmerFuncs import DimmerFuncs
